interface/requirements.txt
kafka-python==2.2.6
pandas==2.3.0
streamlit==1.46.1
psycopg2-binary
matplotlib
numpy
